fix: updated naming convention
