+++ .gitignore
@@ -10,6 +10,8 @@ lerna-debug.log*
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
+
+
 # Runtime data
 pids
 *.pid
@@ -75,11 +77,9 @@ web_modules/
 
 # dotenv environment variable files
 .env
-.env.development.local
-.env.test.local
-.env.production.local
-.env.local
-config.env
+dev.env
+prod.env
+staging.env
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
