+++ app.js
@@ -2,14 +2,20 @@ import express from "express";
 import cors from "cors";
 import dotenv from "dotenv";
 import morgan from "morgan";
-import authRouter from "./routes/authRoute.js";
-import destinationRouter from "./routes/DestinationRoutes.js";
-import reviewRouter from "./routes/reviewRoute.js";
-import bookingRouter from "./routes/bookRoute.js";
-import userRouter from "./routes/userRoute.js";
-
+import authRoute from "./routes/AuthRoute.js";
+import destinationRoute from "./routes/DestinationRoute.js";
+import reviewRoute from "./routes/ReviewRoute.js";
+import bookingRoute from "./routes/BookRoute.js";
+import userRoute from "./routes/UserRoute.js";
 import connectDatabase from "./config/database.js";
-dotenv.config({ path: "./config/config.env" });
+
+if (process.env.NODE_ENV === "development") {
+  dotenv.config({ path: "./config/dev.env" });
+} else if (process.env.NODE_ENV === "staging") {
+  dotenv.config({ path: "./config/staging.env" });
+} else if (process.env.NODE_ENV === "production") {
+  dotenv.config({ path: "./config/prod.env" });
+}
 
 connectDatabase();
 
@@ -18,15 +24,18 @@ app.use(cors());
 app.use(express.json());
 app.use(morgan("tiny"));
 
-// set up routes
-app.use("/api/v1", authRouter);
-app.use("/api/v1", destinationRouter);
-app.use("/api/v1", reviewRouter);
-app.use("/api/v1", bookingRouter);
-app.use("/api/v1", userRouter);
+const port = process.env.PORT || 8000;
 
-const server = app.listen(process.env.PORT, () =>
-  console.log(
-    `server started at port ${process.env.PORT} in ${process.env.NODE_ENV}`
-  )
+app.get("/", (req, res) => {
+  res.send({ message: "hello there" });
+});
+
+app.use("/api/v1", authRoute);
+app.use("/api/v1", destinationRoute);
+app.use("/api/v1", reviewRoute);
+app.use("/api/v1", bookingRoute);
+app.use("/api/v1", userRoute);
+
+const server = app.listen(port, () =>
+  console.log(`server started at port ${port} in ${process.env.NODE_ENV}`)
 );
