+++ config/database.js
@@ -1,11 +1,10 @@
 import mongoose from "mongoose";
 
-// this will connect the database and let us login
 mongoose.set("strictQuery", false);
 const connectDatabase = () => {
   mongoose
     .connect(process.env.DB_URL)
-    .then(() => console.log("database conneted"))
+    .then(() => console.log("database conneted", process.env.DB_URL))
     .catch((error) =>
       console.log(`failed to connect [[[[----${error} ----]]]]`)
     );
