+++ controller/booking/CreateBookingController.js
@@ -1,6 +1,6 @@
-import Booking from "../../model/book/bookSchema.js";
-import { Destination } from "../../model/Destination/Destination.js";
-import User from "../../model/User/User.js";
+import Booking from "../../model/book/BookSchema.js";
+import { Destination } from "../../model/destination/DestinationSchema.js";
+import User from "../../model/User/UserSchema.js";
 
 const createBooking = async (req, res) => {
   const { destinationId } = req.params;
