+++ routes/DestinationRoute.js
@@ -1,11 +1,11 @@
 import express from "express";
-import authMiddleware from "../middleware/chekUser.js";
-import checkRole from "../middleware/checkrole.js";
-import createDestination from "../controller/destination/createDestinationController.js";
-import updateDestination from "../controller/destination/updateDestinationController.js";
-import getAllDestination from "../controller/destination/getAlldestinationController.js";
-import getDestination from "../controller/destination/getDestination.js";
-import deleteDestination from "../controller/destination/deleteDestination.js";
+import authMiddleware from "../middleware/CheckTokenMiddleware.js";
+import checkRole from "../middleware/CheckRolesMiddleware.js";
+import createDestination from "../controller/destination/CreateDestinationController.js";
+import updateDestination from "../controller/destination/UpdateDestinationController.js";
+import getAllDestination from "../controller/destination/GetAlldestinationController.js";
+import getDestination from "../controller/destination/GetDestinationController.js";
+import deleteDestination from "../controller/destination/DeleteDestinationController.js";
 
 const router = express.Router();
 import multer from "multer";
